hw9: moved classes and interfaces to separate modules
