+++ homework-09/src/interfaces/cargo-carrier.ts
@@ -1,0 +1,6 @@
+// Інтерфейс для вантажівок
+export interface ICargoCarrier {
+    loadCargo(weight: number): void;
+    unloadCargo(weight: number): void;
+    getCurrentLoad(): number;
+}
