+++ homework-09/src/interfaces/vehicle.ts
@@ -1,0 +1,7 @@
+// Інтерфейс, який визначає основні операції транспортного засобу
+export interface IVehicle {
+    start(): void;
+    stop(): void;
+    getFuelLevel(): number;
+    refuel(abount: number): void;
+}
