+++ homework-09/src/vehicles.ts
@@ -1,112 +1,7 @@
-// Інтерфейс, який визначає основні операції транспортного засобу
-interface IVehicle {
-    start(): void;
-    stop(): void;
-    getFuelLevel(): number;
-    refuel(abount: number): void;
-}
-
-// Інтерфейс для вантажівок
-interface ICargoCarrier {
-    loadCargo(weight: number): void;
-    unloadCargo(weight: number): void;
-    getCurrentLoad(): number;
-}
-
-// Абстрактний клас, який реалізує загальний функціонал транспортного засобу
-abstract class Vehicle implements IVehicle {
-    protected running= false;
-    protected fuelLevel = 0;
-    protected readonly maxFuel: number;
-    protected readonly model: string;
-
-    public constructor(model: string, maxFuel: number) {
-        this.model = model;
-        this.maxFuel = maxFuel;
-    }
-
-    // Реалізація методів інтерфейсу IVehicle
-    public start(): void {
-        if (this.fuelLevel > 0) {
-            this.running = true;
-            console.log(`${this.model} has started`);
-        } else {
-            console.log(`${this.model} cant't start - no fuel`);
-        }
-    }
-
-    public stop(): void {
-        this.running = false;
-        console.log(`${this.model} has stopped`);
-    }
-
-    public getFuelLevel(): number {
-        return this.fuelLevel;
-    }
-
-    public refuel(amount: number): void {
-        if (amount < 0) {
-            throw new Error('Cannot refuel negative amount');
-        }
-        const refueledLevel = (this.fuelLevel + amount > this.maxFuel)
-            ? this.maxFuel
-            : this.fuelLevel + amount;
-        this.fuelLevel = refueledLevel;
-        console.log(`${this.model} refueled to ${this.fuelLevel} liters`);
-    }
-
-    // Захищений метод для дочірніх класів
-    protected checkEngineStatus(): boolean {
-        return this.running;
-    }
-}
-
-// Клас, який представляє авто
-class Car extends Vehicle {
-    private readonly numDoors: number;
-
-    public constructor(model: string, maxFuel: number, numDoors: number) {
-        super(model, maxFuel);
-        this.numDoors = numDoors;
-    }
-
-    public honk(): void {
-        console.log('Beep beep! :)');
-    }
-}
-
-// Конкретний клас, який представляє вантажівку
-class Truck extends Vehicle implements ICargoCarrier {
-    private currentLoad = 0;
-    private readonly maxLoad: number;
-
-    public constructor(model: string, maxFuel: number, maxLoad: number) {
-        super(model, maxFuel);
-        this.maxLoad = maxLoad;
-    }
-
-    public loadCargo(weight: number): void {
-        if (this.currentLoad + weight <= this.maxLoad) {
-            this.currentLoad += weight;
-            console.log(`Loaded ${weight}kg. Current load: ${this.currentLoad}kg`);
-        } else {
-            throw new Error('Exceeds maximum load capacity');
-        }
-    }
-
-    public unloadCargo(weight: number): void {
-        if (this.currentLoad - weight >= 0) {
-            this.currentLoad -= weight;
-            console.log(`Unloaded ${weight}kg, Current load: ${this.currentLoad}kg`);
-        } else {
-            throw new Error('Cannot unload more than current load');
-        }
-    }
-
-    public getCurrentLoad(): number {
-        return this.currentLoad;
-    }
-}
+import { IVehicle } from '../src/interfaces/vehicle';
+import { ICargoCarrier } from '../src/interfaces/cargo-carrier';
+import { Car } from '../src/classes/car';
+import { Truck } from '../src/classes/truck';
 
 // Функція для тестування автівок
 function testVehicle(vehicle: IVehicle): void {
